+++ src/main/java/com/example/newsfeed/service/UserService.java
@@ -243,18 +243,18 @@ public class UserService {
 
     public List<User> findSuggestion(String username) {
         User byEmail = findByEmail(username);
-        List<Integer> numbers = new ArrayList<>();
+        List<Long> numbers = new ArrayList<>();
 
         List<User> all = userRepository.findAll();
         int size = all.size();
 
-        for (int i = 1; i < size; i++) {
+        for (long i = 1; i < size; i++) {
             if (i==byEmail.getId()) {
                 numbers.add(i);
             }
         }
         Collections.shuffle(numbers);
-        List<Integer> list = numbers.subList(0, 2);
+        List<Long> list = numbers.subList(0, 2);
         return userRepository.findAllByIdIn(list);
     }
 }
